# Pre-sort validator set