--- a/snow/validators/traced_state.go
+++ b/snow/validators/traced_state.go
@@ -64,7 +64,7 @@ func (s *tracedState) GetValidatorSet(
 	ctx context.Context,
 	height uint64,
 	subnetID ids.ID,
-) (map[ids.NodeID]*GetValidatorOutput, error) {
+) ([]*GetValidatorOutput, error) {
 	ctx, span := s.tracer.Start(ctx, s.getValidatorSetTag, oteltrace.WithAttributes(
 		attribute.Int64("height", int64(height)),
 		attribute.Stringer("subnetID", subnetID),