## Why this should be merged

Performance optimization for `GetCanonicalValidatorSet` where we identified two main points of CPU churn, sorting the validator set to make it canonical and deserialization of the bls keys.

This PR aims at maintaining the validator set in a sorted order beforehand so we don't have to re-sort it in the `warp` package.

## How this works

Adds a sorted view of validators by subnet by maintaining them in a map of trees, sorted by NodeID.

Currently, we return an unordered map of validators by fetching the current validator set, and applying a set of diffs up to the height requested.

In this PR we do the following to optimize the existing workflow:

1. Get the current validator set - now the validator set for each subnet is stored in a tree ordered by NodeID. We can iterate over a subnet's tree to get this in linear time.
2. Apply a set of diffs - the diffs are applied as done previously but are stored in a list and sorted afterwards. For requests that require us to go deep into the P-chain this will be the major source of churn, but we expect the `warp` package to hit somewhere close to dip and generally we expect the amount of validators joining/leaving to be relatively small so this sort operation should be generally not too expensive.
3. Merge the two sorted lists - this is done in linear time similar to merge sort

## How this was tested