@joshua-kim 2023-05-22
Initial results running on https://github.com/ava-labs/avalanchego/pull/1467 benchmark tests:

Before:
```
BenchmarkGetCanonicalValidatorSetBySize
BenchmarkGetCanonicalValidatorSetBySize/0
BenchmarkGetCanonicalValidatorSetBySize/0-10   	     259	   4724836 ns/op
BenchmarkGetCanonicalValidatorSetBySize/1
BenchmarkGetCanonicalValidatorSetBySize/1-10   	     258	   4565765 ns/op
BenchmarkGetCanonicalValidatorSetBySize/10
BenchmarkGetCanonicalValidatorSetBySize/10-10  	     271	   4418596 ns/op
BenchmarkGetCanonicalValidatorSetBySize/100
BenchmarkGetCanonicalValidatorSetBySize/100-10 	     270	   4371188 ns/op
BenchmarkGetCanonicalValidatorSetBySize/1000
BenchmarkGetCanonicalValidatorSetBySize/1000-10         	     273	   4456230 ns/op
BenchmarkGetCanonicalValidatorSetBySize/10000
BenchmarkGetCanonicalValidatorSetBySize/10000-10        	     268	   4392789 ns/op
BenchmarkGetCanonicalValidatorSetByDepth
BenchmarkGetCanonicalValidatorSetByDepth/0
BenchmarkGetCanonicalValidatorSetByDepth/0-10           	   28999	     35512 ns/op
BenchmarkGetCanonicalValidatorSetByDepth/1
BenchmarkGetCanonicalValidatorSetByDepth/1-10           	   37110	     31961 ns/op
BenchmarkGetCanonicalValidatorSetByDepth/2
BenchmarkGetCanonicalValidatorSetByDepth/2-10           	   43402	     27497 ns/op
BenchmarkGetCanonicalValidatorSetByDepth/3
BenchmarkGetCanonicalValidatorSetByDepth/3-10           	   48517	     24286 ns/op
BenchmarkGetCanonicalValidatorSetByDepth/4
BenchmarkGetCanonicalValidatorSetByDepth/4-10           	   56355	     20715 ns/op
BenchmarkGetCanonicalValidatorSetByDepth/5
BenchmarkGetCanonicalValidatorSetByDepth/5-10           	   68905	     17252 ns/op
BenchmarkGetCanonicalValidatorSetByDepth/6
BenchmarkGetCanonicalValidatorSetByDepth/6-10           	   85176	     13829 ns/op
BenchmarkGetCanonicalValidatorSetByDepth/7
BenchmarkGetCanonicalValidatorSetByDepth/7-10           	  110547	     10769 ns/op
BenchmarkGetCanonicalValidatorSetByDepth/8
BenchmarkGetCanonicalValidatorSetByDepth/8-10           	  157425	      7463 ns/op
BenchmarkGetCanonicalValidatorSetByDepth/9
BenchmarkGetCanonicalValidatorSetByDepth/9-10           	  278445	      4103 ns/op
PASS
````

After
```
BenchmarkGetCanonicalValidatorSetBySize
BenchmarkGetCanonicalValidatorSetBySize/0
BenchmarkGetCanonicalValidatorSetBySize/0-10   	     292	   4072039 ns/op
BenchmarkGetCanonicalValidatorSetBySize/1
BenchmarkGetCanonicalValidatorSetBySize/1-10   	     295	   4143954 ns/op
BenchmarkGetCanonicalValidatorSetBySize/10
BenchmarkGetCanonicalValidatorSetBySize/10-10  	     280	   4065911 ns/op
BenchmarkGetCanonicalValidatorSetBySize/100
BenchmarkGetCanonicalValidatorSetBySize/100-10 	     292	   4060709 ns/op
BenchmarkGetCanonicalValidatorSetBySize/1000
BenchmarkGetCanonicalValidatorSetBySize/1000-10         	     294	   4110517 ns/op
BenchmarkGetCanonicalValidatorSetBySize/10000
BenchmarkGetCanonicalValidatorSetBySize/10000-10        	     292	   4064871 ns/op
BenchmarkGetCanonicalValidatorSetByDepth
BenchmarkGetCanonicalValidatorSetByDepth/0
BenchmarkGetCanonicalValidatorSetByDepth/0-10           	   36110	     33071 ns/op
BenchmarkGetCanonicalValidatorSetByDepth/1
BenchmarkGetCanonicalValidatorSetByDepth/1-10           	   40346	     29742 ns/op
BenchmarkGetCanonicalValidatorSetByDepth/2
BenchmarkGetCanonicalValidatorSetByDepth/2-10           	   45097	     26533 ns/op
BenchmarkGetCanonicalValidatorSetByDepth/3
BenchmarkGetCanonicalValidatorSetByDepth/3-10           	   50918	     23189 ns/op
BenchmarkGetCanonicalValidatorSetByDepth/4
BenchmarkGetCanonicalValidatorSetByDepth/4-10           	   59392	     20053 ns/op
BenchmarkGetCanonicalValidatorSetByDepth/5
BenchmarkGetCanonicalValidatorSetByDepth/5-10           	   70779	     16672 ns/op
BenchmarkGetCanonicalValidatorSetByDepth/6
BenchmarkGetCanonicalValidatorSetByDepth/6-10           	   88676	     13437 ns/op
BenchmarkGetCanonicalValidatorSetByDepth/7
BenchmarkGetCanonicalValidatorSetByDepth/7-10           	  114945	     10397 ns/op
BenchmarkGetCanonicalValidatorSetByDepth/8
BenchmarkGetCanonicalValidatorSetByDepth/8-10           	  164151	      7080 ns/op
BenchmarkGetCanonicalValidatorSetByDepth/9
BenchmarkGetCanonicalValidatorSetByDepth/9-10           	  302568	      3826 ns/op
PASS
```